small refactor in ConsoleImput.cpp
if a python script returns None, the client interprets that as an error and shows the first parameter to the user.
if the command entered had no parameters, the client will seg fault. Now, the index 0 is shown to the user.

Lastly, if the script returns True, the client does not interpret that as en error.
this is simply to allow scripts to show some message to the user and
exit gracefully (purely aesthetics, not a bug)